feat: async init indexdb

+++ src/cache-provider.ts
@@ -10,13 +10,13 @@ type Key = string
 /**
  * Cache provider factory
  */
-export default async function createCacheProvider<Data = any, Error = any>({
+export default function createCacheProvider<Data = any, Error = any>({
   dbName,
   storeName,
   storageHandler = simpleStorageHandler,
   version = 1,
   onError = () => {},
-}: Config): Promise<CacheProvider> {
+}: Config): CacheProvider {
   type Cache = SWRCache<Data>
   type State = SWRState<Data, Error>
 
@@ -26,7 +26,9 @@ export default async function createCacheProvider<Data = any, Error = any>({
   // Initialize database
   let db: IDBPDatabase
 
-  try {
+  let initLock = true // Lock the database operation until the database finishes initializing
+
+  async function initDb() {
     db = await openDB(dbName, version, {
       upgrade (upgradeDb, oldVersion) {
         if (!oldVersion) {
@@ -54,11 +56,10 @@ export default async function createCacheProvider<Data = any, Error = any>({
 
       cursor = await cursor.continue()
     }
-  } catch (error) {
-    // Use fallback
-    onError(error)
-    return (): Cache => map
   }
+  initDb().then(() => {
+    initLock = false
+   })
 
   /**
    * SWR Cache provider API
@@ -73,7 +74,7 @@ export default async function createCacheProvider<Data = any, Error = any>({
     set: (key: Key, value: State): void => {
       map.set(key, value)
 
-      if (isFetchInfo(value)) {
+      if (isFetchInfo(value) || initLock) {
         return
       }
 
@@ -92,6 +93,9 @@ export default async function createCacheProvider<Data = any, Error = any>({
      */
     delete: (key: Key): void => {
       if (map.delete(key)) {
+        if (initLock) {
+          return
+        }
         db.delete(storeName, key)
           .catch(onError)
       }
